requirements.txt
numpy
paho-mqtt
fastapi
uvicorn
pymongo